refactor: 시간 기능의 Service 계층 분리
