+++ src/test/java/roomescape/api/TimeControllerTest.java
@@ -11,10 +11,32 @@ import org.springframework.test.context.jdbc.Sql;
 import java.util.HashMap;
 import java.util.Map;
 
+import static org.hamcrest.Matchers.is;
+
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @Sql("classpath:init.sql")
-class TimeCommandControllerTest {
+class TimeControllerTest {
+
+    @Test
+    @DisplayName("시간 목록을 조회할 수 있다.")
+    void findTimes() {
+        Map<String, String> params = new HashMap<>();
+        params.put("time", "10:00");
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/times")
+                .then().log().all();
+
+        RestAssured.given().log().all()
+                .when().get("/times")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(1));
+    }
+
     @Test
     @DisplayName("시간을 추가할 수 있다.")
     void addTimes() {
